+++ benches/benchmarks/with_inputs.rs
@@ -9,7 +9,7 @@ fn from_elem(c: &mut Criterion) {
 
     let mut group = c.benchmark_group("from_elem");
     for size in [KB, 2 * KB, 4 * KB, 8 * KB, 16 * KB].iter() {
-        group.throughput(Throughput::Bytes(*size as u32));
+        group.throughput(Throughput::Bytes(*size as u64));
         group.bench_with_input(BenchmarkId::from_parameter(size), size, |b, &size| {
             b.iter(|| iter::repeat(0u8).take(size).collect::<Vec<_>>());
         });
